+++ package.json
@@ -21,7 +21,7 @@
     }
   },
   "lint-staged": {
-    "*.(js|jsx|ts|tsx|css|json)": [
+    "*.(js|jsx|ts|tsx|json)": [
       "npm run lint:fix",
       "npm run format:fix"
     ]
@@ -34,6 +34,7 @@
   "devDependencies": {
     "@jest/globals": "^29.7.0",
     "@testing-library/react": "^14.1.0",
+    "@types/jest": "^29.5.8",
     "@types/react": "^18.2.15",
     "@types/react-dom": "^18.2.7",
     "@typescript-eslint/eslint-plugin": "^6.9.0",
@@ -48,6 +49,7 @@
     "gh-pages": "^6.0.0",
     "husky": "^8.0.3",
     "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "lint-staged": "^15.1.0",
     "prettier": "^3.0.3",
     "ts-jest": "^29.1.1",
