+++ src/components/header/Header.test.tsx
@@ -1,0 +1,8 @@
+import { render } from "@testing-library/react";
+import Header from "./Header";
+
+describe("Header component", () => {
+  test("it renders", () => {
+    render(<Header />);
+  });
+});
